[Done] Calibration example
